services/product/v2/pyproject.toml
[tool.poetry]
name = "product"
version = "2.0.0"
description = "product management service"
authors = ["T.H."]
license = "MIT"
readme = "README.md"
packages = [{include="product", from="./src"}, {include="settings", from="."}, {include="tests", from="."}]

[tool.poetry.dependencies]
python = "3.12"
granian = "1.5.2"
blacksheep = "2.0.6"
pydantic = "2.9.2"
elasticsearch-curator = {git = "https://github.com/metalalive/curator.git",  branch = "5.5"}
aiohttp = "~3.10.0"
ecommerce-common = {path = "../../common/python"}
my-c-extension-lib = {path = "../../common/python/c_exts"}

[tool.poetry.group.test.dependencies]
pytest = ">=8.2.0"
pytest-asyncio = ">=0.24.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
# tracking relevant issue:
# https://github.com/pytest-dev/pytest-asyncio/issues/924
